+++ src/features/events/events.model.js
@@ -1,4 +1,6 @@
-module.exports = (sequelize, DataTypes) => {
+const sequelize = require("../../shared/database/index");
+const DataTypes = require("sequelize");
+module.exports = () => {
   const EventLog = sequelize.define(
     "EventLog",
     {
