+++ src/features/student/student.model.js
@@ -1,4 +1,7 @@
-module.exports = (sequelize, DataTypes) => {
+const sequelize = require("../../shared/database/index");
+const DataTypes = require("sequelize");
+
+module.exports = () => {
   const Student = sequelize.define(
     "Student",
     {
