+++ src/features/user/user.model.js
@@ -1,6 +1,7 @@
 const userAuthPlugin = require("./userAuth.plugin");
-
-module.exports = (sequelize, DataTypes) => {
+const sequelize = require("../../shared/database/index");
+const DataTypes = require("sequelize");
+module.exports = () => {
   const User = sequelize.define(
     "User",
     {
